requirements.txt
pytest==7.1.1
mock
PyYAML==6.0
kafka-schema-registry==0.1.2
black
tokenize-rt
